requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
torch
torchvision
pillow==10.3.0
python-multipart==0.0.9
numpy==1.26.4
opencv-python==4.9.0.80
tqdm==4.66.2
pydantic-settings==2.2.1
